Lab6 done

+++ src/main/java/org/example/post/service/PostService.java
@@ -1,15 +1,19 @@
 package org.example.post.service;
 
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ejb.EJBAccessException;
 import jakarta.ejb.LocalBean;
 import jakarta.ejb.Stateless;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.security.enterprise.SecurityContext;
 import jakarta.transaction.Transactional;
 import lombok.NoArgsConstructor;
 import org.example.post.entity.Post;
 import org.example.category.repository.api.CategoryRepository;
 import org.example.post.repository.api.PostRepository;
 import org.example.user.entity.User;
+import org.example.user.entity.UserRoles;
 import org.example.user.repository.api.UserRepository;
 
 import java.util.List;
@@ -39,16 +43,20 @@ public class PostService {
      */
     private final UserRepository userRepository;
 
+    private final SecurityContext securityContext;
+
     /**
      * @param postRepository  repository for post entity
      * @param categoryRepository repository for category entity
      * @param userRepository repository for user entity
      */
     @Inject
-    public PostService(PostRepository postRepository, CategoryRepository categoryRepository, UserRepository userRepository) {
+    public PostService(PostRepository postRepository, CategoryRepository categoryRepository, UserRepository userRepository,
+    @SuppressWarnings("CdiInjectionPointsInspection") SecurityContext securityContext) {
         this.postRepository = postRepository;
         this.categoryRepository = categoryRepository;
         this.userRepository = userRepository;
+        this.securityContext = securityContext;
     }
 
     /**
@@ -57,13 +65,20 @@ public class PostService {
      * @param id post's id
      * @return container with post
      */
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
     public Optional<Post> find(UUID id) {
         return postRepository.find(id);
+    }
+
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
+    public Optional<Post> find(User user,UUID id) {
+        return postRepository.findByIdAndUser(id,user);
     }
 
     /**
      * @return all available posts
      */
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
     public List<Post> findAll() {
         return postRepository.findAll();
     }
@@ -72,8 +87,33 @@ public class PostService {
      * @param user existing user, post's owner
      * @return all available posts of the selected user
      */
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
     public List<Post> findAll(User user) {
         return postRepository.findAllByUser(user);
+    }
+
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
+    public List<Post> findAllForCallerPrincipal() {
+        if(securityContext.isCallerInRole(UserRoles.ADMIN)){
+            return postRepository.findAll();
+        }
+        User user = userRepository.findByName(securityContext.getCallerPrincipal().getName()).orElseThrow(IllegalStateException::new);
+        return findAll(user);
+    }
+
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
+    public Optional<Post> findForCallerPrincipal(UUID id) {
+        Post post = find(id).orElseThrow(IllegalStateException::new);
+
+        checkAdminRoleOrOwner(Optional.ofNullable(post));
+
+        return Optional.ofNullable(post);
+
+//        if(securityContext.isCallerInRole(UserRoles.ADMIN)){
+//            return find(id);
+//        }
+//        User user = userRepository.findByName(securityContext.getCallerPrincipal().getName()).orElseThrow(IllegalStateException::new);
+//        return find(user,id);
     }
 
     /**
@@ -81,8 +121,15 @@ public class PostService {
      *
      * @param post new post
      */
-
+    @RolesAllowed(UserRoles.ADMIN)
     public void create(Post post) {
+        postRepository.create(post);
+    }
+
+    @RolesAllowed(UserRoles.USER)
+    public void createForCallerPrincipal(Post post) {
+        User user = userRepository.findByName(securityContext.getCallerPrincipal().getName()).orElseThrow(IllegalStateException::new);
+        post.setUser(user);
         postRepository.create(post);
     }
 
@@ -92,7 +139,9 @@ public class PostService {
      * @param post post to be updated
      */
 
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
     public void update(Post post) {
+        checkAdminRoleOrOwner(postRepository.find(post.getId()));
         postRepository.update(post);
     }
 
@@ -102,7 +151,9 @@ public class PostService {
      * @param id existing post's id to be deleted
      */
 
+    @RolesAllowed({UserRoles.USER,UserRoles.ADMIN})
     public void delete(UUID id) {
+        checkAdminRoleOrOwner(postRepository.find(id));
         postRepository.delete(postRepository.find(id).orElseThrow());
     }
 
@@ -118,13 +169,32 @@ public class PostService {
 //        });
 //    } //FIXME: Avatar
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<List<Post>> findAllByCategory(UUID id) {
         return categoryRepository.find(id)
                 .map(postRepository::findAllByCategory);
     }
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<List<Post>> findAllByUser(UUID id) {
         return userRepository.find(id)
                 .map(postRepository::findAllByUser);
     }
+
+    /**
+     * @param post post to be checked
+     * @throws EJBAccessException when caller principal has no admin role and is not character's owner
+     */
+    private void checkAdminRoleOrOwner(Optional<Post> post) throws EJBAccessException {
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return;
+        }
+        if (securityContext.isCallerInRole(UserRoles.USER)
+                && post.isPresent()
+                && post.get().getUser().getName().equals(securityContext.getCallerPrincipal().getName())) {
+            return;
+        }
+        throw new EJBAccessException("Caller not authorized.");
+    }
+
 }
